+++ src/logger.ts
@@ -3,22 +3,24 @@ import os from 'os';
 import yaml from 'js-yaml';
 import AWS from 'aws-sdk';
 import vscode from "vscode";
+import macaddress from "macaddress";
 
 import { PutLogEventsRequest } from 'aws-sdk/clients/cloudwatchlogs';
 import {
-	AWS_REGION,
-	CLIENT_LOGS_GROUP_NAME,
 	DIFF_SOURCE,
 	LOG_AFTER_X_TIMES
 } from './constants';
 import { readYML, isEmpty } from './utils/common';
-import { generateSettings } from "./settings";
+import { generateSettings, LOGS_METADATA, PLUGIN_USER } from "./settings";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 const VERSION = vscode.extensions.getExtension('codesync.codesync').packageJSON.version;
 
 let cloudwatchlogs = <AWS.CloudWatchLogs>{};
+let macAddress = "";
+macaddress.one().then(mac => macAddress = mac);
+
 const logErrorMsgTypes = {
 	CRITICAL: "CRITICAL",
 	ERROR: "ERROR",
@@ -36,78 +38,91 @@ export class CodeSyncLogger {
 	  CRITICAL: Errors that are blocking for the normal operation of the plugin and should be fixed immediately.
 	*/
 
-	static debug (msg: string, additionalMsg="", userEmail?: string, retryCount=0) {
-		putLogEvent(msg, logErrorMsgTypes.DEBUG, additionalMsg, userEmail, retryCount);
+	static debug (msg: string, additionalMsg="", logStream?: string, retryCount=0) {
+		putLogEvent(msg, logErrorMsgTypes.DEBUG, additionalMsg, logStream, retryCount);
 	}
 
-	static info (msg: string, additionalMsg="", userEmail?: string, retryCount=0) {
-		putLogEvent(msg, logErrorMsgTypes.INFO, additionalMsg, userEmail, retryCount);
+	static info (msg: string, additionalMsg="", logStream?: string, retryCount=0) {
+		putLogEvent(msg, logErrorMsgTypes.INFO, additionalMsg, logStream, retryCount);
 	}
 
-	static warning (msg: string, additionalMsg="", userEmail?: string, retryCount=0) {
-		putLogEvent(msg, logErrorMsgTypes.WARNING, additionalMsg, userEmail, retryCount);
+	static warning (msg: string, additionalMsg="", logStream?: string, retryCount=0) {
+		putLogEvent(msg, logErrorMsgTypes.WARNING, additionalMsg, logStream, retryCount);
 	}
 
-	static error (msg: string, additionalMsg="", userEmail?: string, retryCount=0) {
-		putLogEvent(msg, logErrorMsgTypes.ERROR, additionalMsg, userEmail, retryCount);
+	static error (msg: string, additionalMsg="", logStream?: string, retryCount=0) {
+		putLogEvent(msg, logErrorMsgTypes.ERROR, additionalMsg, logStream, retryCount);
 	}
 
-	static critical (msg: string, additionalMsg="", userEmail?: string, retryCount=0) {
-		putLogEvent(msg, logErrorMsgTypes.CRITICAL, additionalMsg, userEmail, retryCount);
+	static critical (msg: string, additionalMsg="", logStream?: string, retryCount=0) {
+		putLogEvent(msg, logErrorMsgTypes.CRITICAL, additionalMsg, logStream, retryCount);
 	}
 
 }
 
-const putLogEvent = (msg: string, eventType: string, additionalMsg="", userEmail?: string, retryCount=0) => {
+const putLogEvent = (msg: string, eventType: string, additionalMsg="", logStream?: string, retryCount=0) => {
 	let eventMsg = msg;
 	if (additionalMsg) {
 		eventMsg = `${msg}, ${additionalMsg}`;
 	}
 	console.log(eventMsg);
+
+	let email = "";
+	let accessKey = "";
+	let secretKey = "";
+
 	const settings = generateSettings();
-	if (!fs.existsSync(settings.USER_PATH)) return;
+
+	if (!fs.existsSync(settings.USER_PATH) || !fs.existsSync(settings.SEQUENCE_TOKEN_PATH)) return;
+
 	const users = readYML(settings.USER_PATH);
 	const sequenceTokenConfig = readYML(settings.SEQUENCE_TOKEN_PATH);
-	let accessKey = '';
-	let secretKey = '';
-	let sequenceToken = '';
-	let email = '';
 
-	if (userEmail && userEmail in users) {
-		const user = users[userEmail];
-		email = userEmail;
-		accessKey = user.access_key;
-		secretKey = user.secret_key;
-		sequenceToken = sequenceTokenConfig[userEmail];
+	if (logStream) {
+		const user = users[logStream];
+		if (user && user.is_active) {
+			email = logStream;
+			accessKey = user.access_key;
+			secretKey = user.secret_key;
+		}
 	} else {
-		Object.keys(users).forEach(function (_email, index) {
-			if (index === 0) {
-				email = _email;
-				const user = users[email];
-				accessKey = user.access_key;
-				secretKey = user.secret_key;
-				sequenceToken = sequenceTokenConfig[email];
-			}
-		});
+		const activeUserEmail = Object.keys(users).filter(_email => users[_email].is_active)[0];
+		if (activeUserEmail) {
+			email = activeUserEmail;
+			accessKey = users[activeUserEmail].access_key;
+			secretKey = users[activeUserEmail].secret_key;
+		}
 	}
 
+	// Set default user for logging
 	if (!(accessKey && secretKey && email)) {
-		return;
+		email = PLUGIN_USER.logStream;
+		const pluginUser = users[email];
+		if (!pluginUser) return;
+		accessKey = pluginUser.access_key;
+		secretKey = pluginUser.secret_key;
 	}
+	
+	const sequenceToken = email in sequenceTokenConfig ? sequenceTokenConfig[email] : "";
+
 	if (isEmpty(cloudwatchlogs)) {
 		cloudwatchlogs = new AWS.CloudWatchLogs({
 			accessKeyId: accessKey,
 			secretAccessKey: secretKey,
-			region: AWS_REGION
+			region: LOGS_METADATA.AWS_REGION
 		});
 	}
+
+	const logGroupName = LOGS_METADATA.GROUP;
+	const logStreamName = email;
 
 	const CWEventMsg = {
 		msg: eventMsg,
 		type: eventType,
 		source: DIFF_SOURCE,
 		version: VERSION,
-		platform: os.platform()
+		platform: os.platform(),
+		mac_address: macAddress
 	};
 	const logEvents = [ /* required */
 		{
@@ -115,8 +130,6 @@ const putLogEvent = (msg: string, eventType: string, additionalMsg="", userEmail
 			timestamp: new Date().getTime() /* required */
 		}
 	];
-	const logGroupName = CLIENT_LOGS_GROUP_NAME;
-	const logStreamName = email;
 
 	const params = <PutLogEventsRequest>{
 		logEvents,
@@ -141,7 +154,7 @@ const putLogEvent = (msg: string, eventType: string, additionalMsg="", userEmail
 		The next batch can be sent with sequenceToken: 49615429905286623782064446503967477603282951356289123634
 		*/
 		const errString = err.toString();
-		if (errString.substr('DataAlreadyAcceptedException') || errString.substr('InvalidSequenceTokenException')) {
+		if (errString.includes('DataAlreadyAcceptedException') || errString.includes('InvalidSequenceTokenException')) {
 			const matches = errString.match(/(\d+)/);
 			if (matches[0]) {
 				sequenceTokenConfig[email] = matches[0];
@@ -149,16 +162,16 @@ const putLogEvent = (msg: string, eventType: string, additionalMsg="", userEmail
 				if (retryCount) {
 					if (retryCount < 10) {
 						retryCount += 1;
-						putLogEvent(msg, eventType, email, additionalMsg, retryCount);
+						putLogEvent(msg, eventType, additionalMsg, email, retryCount);
 					}
 				} else {
-					putLogEvent(msg, eventType, email, additionalMsg, 1);
+					putLogEvent(msg, eventType, additionalMsg, email, 1);
 				}
 			} else {
-				console.log(err, err.stack);
+				console.log(`Logging failed: ${errString}`, err.stack);
 			}
 		} else {
-			console.log(err, err.stack);
+			console.log(`Failed to log: ${errString}`);
 		}
 	});
 };