+++ src/utils/api_utils.ts
@@ -1,6 +1,7 @@
 import fetch from "node-fetch";
 
 import { API_ROUTES } from "../constants";
+import { PLUGIN_USER } from "../settings";
 
 export const checkServerDown = async () => {
 	let isDown = false;
@@ -62,6 +63,49 @@ export const createUserWithApi = async (accessToken: string) => {
 
 	return {
 		email,
+		error
+	};
+};
+
+export const getPluginUser = async () => {
+	let error = "";
+	const response = await fetch(PLUGIN_USER.url)
+		.then(res => res.json())
+		.then(json => json)
+		.catch(err => error = err);
+
+	if (response.error) {
+		error = response.error.message;
+	}
+
+	return {
+		user: {...response},
+		error
+	};
+};
+
+
+export const getTeamActivity = async (accessToken: string) => {
+	let error = "";
+	let activities = <any>[];
+	const response = await fetch(
+		API_ROUTES.TEAM_ACTIVITY, {
+			headers: {
+				'Content-Type': 'application/json',
+				'Authorization': `Basic ${accessToken}`
+			}
+	})
+		.then(res => res.json())
+		.then(json => json)
+		.catch(err => error = err);
+	if (response.error) {
+		error = response.error.message;
+	} else {
+		activities = response.activities;
+	}
+
+	return {
+		activities,
 		error
 	};
 };